% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ayce_delay_adjustment.R
\name{apply_delay_adjustment}
\alias{apply_delay_adjustment}
\title{Delay Adjustment : Delay Adjustment (Version Rapide)}
\usage{
apply_delay_adjustment(
  data,
  delay_range = -25:25,
  n_iterations = 5,
  noise_level = NULL,
  value_col = "Flow",
  sample_fraction = NULL,
  method = NULL,
  coarse_step = 2,
  max_points = 10000,
  bandwidth = 30
)
}
\arguments{
\item{data}{Tibble avec donnees de rendement (X, Y, Flow, GPS_Time, Interval)}

\item{delay_range}{Plage de delais a tester (defaut -25:25 secondes)}

\item{value_col}{Nom de la colonne de valeurs a analyser}

\item{coarse_step}{Pas pour la recherche grossiere (defaut 2)}

\item{max_points}{Nombre maximum de points pour l'analyse}

\item{bandwidth}{Distance de bande passante pour Moran (defaut 30m)}
}
\value{
Liste avec optimal_delay, score_values et stability_metrics
}
\description{
Determine automatiquement le delai optimal entre le flux et la position GPS
en utilisant une recherche en deux etapes (grossiere + fine) avec rgeoda.
}
\examples{
\dontrun{
result <- apply_delay_adjustment(data, delay_range = -25:25)
}
}
