% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_yield.R
\name{apply_moisture_delay}
\alias{apply_moisture_delay}
\title{Appliquer la correction de delai d'humidite}
\usage{
apply_moisture_delay(data, delay = 15, direction = "forward")
}
\arguments{
\item{data}{Tibble avec donnees de rendement}

\item{delay}{Nombre d'observations a decaler}

\item{direction}{Direction du decalage}
}
\value{
Tibble avec valeurs d'humidite corrigees
}
\description{
Cette fonction compense le délai entre la mesure d'humidité et la position GPS.
}
