% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ayce_pcdi.R
\name{apply_position_filter}
\alias{apply_position_filter}
\title{Appliquer le filtre de position (POS)}
\usage{
apply_position_filter(data, thresholds)
}
\arguments{
\item{data}{Tibble avec coordonnees X, Y}

\item{thresholds}{Liste des seuils de position}
}
\value{
Tibble filtre
}
\description{
Elimine les flyers GPS en verifiant que les points sont dans l'enveloppe
inter-quantile etendue du champ.
}
