% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_yield_fast.R
\name{apply_yield_filters}
\alias{apply_yield_filters}
\title{Application des filtres de rendement}
\usage{
apply_yield_filters(preprocessed_data, params = list(), polygon = TRUE)
}
\arguments{
\item{preprocessed_data}{Données pré-traitées}

\item{params}{Liste des paramètres des filtres}

\item{polygon}{Booléen pour création de polygones}
}
\value{
Données filtrées
}
\description{
Cette fonction applique les filtres qui peuvent être modifiés
sans recalculer le pré-traitement : header, GPS, vitesse,
plage de rendement, humidité, etc.
}
