% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ayce_pcdi.R
\name{calculate_auto_thresholds}
\alias{calculate_auto_thresholds}
\title{Calculer les seuils automatiques (methode quantiles-IQR)}
\usage{
calculate_auto_thresholds(
  data,
  yllim = 0.05,
  yulim = 0.95,
  yscale = 1.5,
  vllim = 0.02,
  vulim = 0.98,
  vscale = 1.5,
  minv_abs = 0.5,
  miny_abs = 0,
  gbuffer = 100
)
}
\arguments{
\item{data}{Tibble avec donnees de rendement}

\item{yllim}{Limite quantile basse (defaut 0.05)}

\item{yulim}{Limite quantile haute (defaut 0.95)}

\item{yscale}{Facteur d'extension IQR (defaut 1.5)}

\item{vllim}{Limite quantile basse vitesse (defaut 0.02)}

\item{vulim}{Limite quantile haute vitesse (defaut 0.98)}

\item{vscale}{Facteur d'extension IQR vitesse (defaut 1.5)}

\item{minv_abs}{Seuil minimal absolu de vitesse (defaut 0.5 m/s)}

\item{miny_abs}{Seuil minimal absolu de rendement (defaut 0)}

\item{gbuffer}{Marge pour le filtre de position en metres (defaut 100)}
}
\value{
Liste avec tous les seuils calcules
}
\description{
Automatise les filtres MINY, MAXY, MINV, MAXV, POS a partir d'analyses
de distributions basees sur les quantiles.
}
