% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yield_conversion.R
\name{convert_flow_to_yield}
\alias{convert_flow_to_yield}
\title{Convertir le flux de grain en rendement (boisseaux/acre)}
\usage{
convert_flow_to_yield(
  data,
  lbs_per_bushel = NULL,
  sqft_per_acre = 43560,
  inches_per_foot = 12
)
}
\arguments{
\item{data}{Tibble avec Flow, Interval, Swath, Distance}

\item{lbs_per_bushel}{LBS par boisseau. Si NULL, auto-detection via GrainType.
Defaut 56 pour le mais. Utiliser 60 pour soja et cereales.}

\item{sqft_per_acre}{Pieds carres par acre (defaut 43560)}

\item{inches_per_foot}{Pouces par pied (defaut 12)}
}
\value{
Donnees avec colonne Yield_buacre
}
\description{
Convertit le flux brut (LBS/SEC) en rendement (boisseaux/acre)
avec la formule :
Rendement (bu/acre) = (Flow x Interval x 43560) / (lbs_per_bushel x Swath_ft x Distance_ft)
}
\details{
Ou :
\itemize{
\item Flow = flux de grain en lbs/sec
\item Interval = intervalle de temps en secondes
\item Swath_ft = largeur de coupe en pieds (Swath_in / 12)
\item Distance_ft = distance parcourue en pieds (Distance_in / 12)
\item 43560 = pieds^2 par acre
\item lbs_per_bushel = lbs par boisseau (selon la culture)
}

Facteurs de conversion standard :
\itemize{
\item Mais : 56 lbs/bu a 15.5\% humidite
\item Soja : 60 lbs/bu a 13\% humidite
\item Ble/cereales : 60 lbs/bu
}
}
\examples{
\dontrun{
# Auto-detection selon la culture
data <- convert_flow_to_yield(data)

# Explicite pour le mais
data <- convert_flow_to_yield(data, lbs_per_bushel = 56)

# Pour soja/cereales
data <- convert_flow_to_yield(data, lbs_per_bushel = 60)
}
}
