% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{create_polygons_from_data}
\alias{create_polygons_from_data}
\title{Creer des polygones rectangulaires a partir de donnees ponctuelles}
\usage{
create_polygons_from_data(data, heading_col = NULL)
}
\arguments{
\item{data}{Data.frame avec colonnes Longitude, Latitude, et soit
(Swath_m, Distance_m) soit (Swath, Distance) en metres}

\item{heading_col}{Nom de la colonne heading (defaut: "Heading" ou "heading")}
}
\value{
Objet SF avec polygones en UTM
}
\description{
Convertit un data.frame avec coordonnees, heading, swath et distance
en polygones rectangulaires en UTM.
}
