% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_diagnostics.R
\name{export_raster}
\alias{export_raster}
\title{Exporter les donnees nettoyees en raster}
\usage{
export_raster(
  data,
  cell_size = 1,
  column_colonne = "Yield_kg_ha",
  fun = mean,
  crs_code = NULL
)
}
\arguments{
\item{data}{Donnees nettoyees (objet sf avec geometrie points ou polygones)}

\item{cell_size}{Taille des cellules en metres (defaut: 1)}

\item{column_colonne}{Nom de la colonne contenant les valeurs a rasteriser}

\item{fun}{Fonction d'agregation (defaut: mean)}

\item{crs_code}{Code EPSG du systeme de coordonnees (defaut: NULL, auto-detecte)}
}
\value{
Objet SpatRaster
}
\description{
Convertit les donnees de rendement nettoyees en raster avec une resolution
specifique et decoupe selon un polygone concave.
}
