% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_basic.R
\name{filter_bounds}
\alias{filter_bounds}
\title{Filtrer selon les limites geographiques}
\usage{
filter_bounds(data, bounds = NULL, coord_type = "latlon")
}
\arguments{
\item{data}{Tibble avec donnees de rendement}

\item{bounds}{Liste avec min/max des coordonnees x et y}

\item{coord_type}{Type de coordonnees : "utm" ou "latlon"}
}
\value{
Tibble filtre dans les limites
}
\description{
Cette fonction filtre les points selon les limites géographiques
du champ (Easting/Northing ou Lat/Lon).
}
