% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_basic.R
\name{filter_dop}
\alias{filter_dop}
\title{Filtrer selon le DOP (Dilution of Precision)}
\usage{
filter_dop(data, max_dop = 10)
}
\arguments{
\item{data}{Tibble avec donnees de rendement}

\item{max_dop}{Valeur maximale acceptable du DOP (defaut 10)}
}
\value{
Tibble filtre avec DOP valide
}
\description{
Cette fonction élimine les points avec un DOP trop élevé
(mauvaise précision GPS).
}
