% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_basic.R
\name{filter_gps_status}
\alias{filter_gps_status}
\title{Filtrer selon le statut GPS}
\usage{
filter_gps_status(data, min_gps_status = 4)
}
\arguments{
\item{data}{Tibble avec donnees de rendement}

\item{min_gps_status}{Statut GPS minimal (defaut 4 = bon)}
}
\value{
Tibble filtre avec GPS valide
}
\description{
Cette fonction filtre les données selon la qualité du signal GPS.
}
