% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_advanced.R
\name{filter_sliding_window}
\alias{filter_sliding_window}
\title{Appliquer le filtre a fenetre glissante}
\usage{
filter_sliding_window(data, window_size = 11, n_std = 3, yield_col = "Flow")
}
\arguments{
\item{data}{Tibble avec donnees de rendement}

\item{window_size}{Taille de la fenetre glissante}

\item{n_std}{Nombre d'ecarts-types pour le seuil}

\item{yield_col}{Nom de la colonne de rendement}
}
\value{
Tibble filtre
}
\description{
Cette fonction applique un filtre à fenêtre glissante pour éliminer
les valeurs aberrantes basées sur les voisins temporels.
}
