% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_motion.R
\name{filter_velocity_jumps}
\alias{filter_velocity_jumps}
\title{Filtre pour changements brusques de vitesse}
\usage{
filter_velocity_jumps(data, max_acceleration = 5, max_deceleration = -8)
}
\arguments{
\item{data}{Tibble avec au minimum X, Y, Interval et orig_row_id}

\item{max_acceleration}{Acceleration maximale autorisee (m/s, defaut: 5)}

\item{max_deceleration}{Deceleration maximale autorisee (m/s, defaut: -8)}
}
\value{
Liste avec data (donnees filtrees) et removed (points supprimes)
}
\description{
Detecte et supprime les points ou il y a une variation tres grande de la vitesse
entre deux points consecutifs. La vitesse est calculee a partir de la distance
et du temps entre les points.
}
