% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_batch_report.R
\name{generate_batch_report_html}
\alias{generate_batch_report_html}
\title{Generate a multi-field HTML report}
\usage{
generate_batch_report_html(file_paths, output_file = NULL, title = NULL)
}
\arguments{
\item{file_paths}{Vector of file paths (GeoJSON)}

\item{output_file}{Output HTML file path}

\item{title}{Report title}
}
\value{
Path to generated HTML file
}
\description{
Generate a multi-field HTML report
}
