% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize_attributes.R
\name{get_secure_password}
\alias{get_secure_password}
\title{Recuperer un mot de passe stocke de maniere securisee}
\usage{
get_secure_password(
  method = "env_var",
  env_var = "YIELD_ANON_PASSWORD",
  keyring_service = "yieldcleanr",
  config_file = NULL
)
}
\arguments{
\item{method}{Methode de stockage utilisee: "env_var", "keyring", "file"}

\item{env_var}{Nom de la variable d'environnement}

\item{keyring_service}{Nom du service keyring}

\item{config_file}{Chemin du fichier de configuration}
}
\value{
Le mot de passe recupere ou NULL si non trouve
}
\description{
Recuperer un mot de passe stocke de maniere securisee
}
