% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{launch_shiny_app}
\alias{launch_shiny_app}
\title{Lancer l'application Shiny de nettoyage des rendements}
\usage{
launch_shiny_app()
}
\value{
Lance l'application Shiny dans le navigateur (invisible)
}
\description{
Lance une application Shiny interactive pour visualiser le nettoyage des
donnees de rendement. L'utilisateur peut importer un fichier, voir les
suppressions par etape, visualiser les cartes et telecharger les resultats.
}
\examples{
\dontrun{
launch_shiny_app()
}
}
