% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{list_fields_from_zip}
\alias{list_fields_from_zip}
\title{Lister les champs disponibles dans un fichier ZIP}
\usage{
list_fields_from_zip(zip_path)
}
\arguments{
\item{zip_path}{Chemin vers le fichier ZIP}
}
\value{
Un tibble avec les informations sur les champs disponibles
}
\description{
Cette fonction liste tous les champs disponibles dans un fichier ZIP
contenant des shapefiles (format John Deere, etc.)
}
