% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_yield_from_zip}
\alias{read_yield_from_zip}
\title{Lire les donnees de rendement depuis un fichier ZIP}
\usage{
read_yield_from_zip(zip_path, field_name)
}
\arguments{
\item{zip_path}{Chemin vers le fichier ZIP}

\item{field_name}{Nom du champ a lire}
}
\value{
Un objet sf avec les donnees de rendement
}
\description{
Cette fonction lit les donnees de rendement depuis un fichier ZIP
contenant des shapefiles (format John Deere, etc.)
}
