% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize_attributes.R
\name{remove_sensitive_attributes}
\alias{remove_sensitive_attributes}
\title{Supprimer les attributs sensibles des donnees de rendement}
\usage{
remove_sensitive_attributes(
  data,
  columns_to_remove = c("Serial", "FieldID", "LoadID", "GPS_Time"),
  generalize_cols = NULL,
  keep_temporal = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{Tibble ou data.frame avec les donnees de rendement}

\item{columns_to_remove}{Vecteur des noms de colonnes a supprimer.
Par defaut: c("Serial", "FieldID", "LoadID", "GPS_Time")}

\item{generalize_cols}{Liste nommee definissant les colonnes a generaliser
au lieu de supprimer. Ex: list(GPS_Time = "hour", Pass = "none")}

\item{keep_temporal}{Si TRUE, conserve une version temporelle relative
(minutes depuis le debut) au lieu de supprimer GPS_Time completement}

\item{verbose}{Si TRUE, affiche des informations sur les colonnes traitees}
}
\value{
Tibble avec les colonnes sensibles supprimees ou generalisees
}
\description{
Cette fonction supprime les colonnes identifiantes et sensibles des donnees
de rendement pour renforcer l'anonymisation. Elle peut egalement generaliser
certaines colonnes (arrondir, categoriser) plutot que de les supprimer completement.
}
\details{
Les colonnes suivantes sont considerees comme sensibles par defaut:
\itemize{
\item Serial: Numero de serie de la machine (identifiant unique)
\item FieldID: Identifiant du champ (peut reveler la propriete)
\item LoadID: Identifiant de la charge (lien avec d'autres donnees)
\item GPS_Time: Horodatage precis (peut etre croise avec d'autres sources)
}

Options de generalisation pour GPS_Time:
\itemize{
\item "none": Supprime completement
\item "hour": Garde uniquement l'heure (sans minutes/secondes)
\item "day": Garde uniquement la date (sans heure)
\item "relative": Convertit en minutes ecoulees depuis le debut
}
}
\examples{
\dontrun{
# Supprimer les colonnes sensibles par defaut
data_clean <- remove_sensitive_attributes(data)

# Supprimer des colonnes specifiques
data_clean <- remove_sensitive_attributes(
  data,
  columns_to_remove = c("Serial", "FieldID", "Variety")
)

# Generaliser plutot que supprimer
data_clean <- remove_sensitive_attributes(
  data,
  generalize_cols = list(GPS_Time = "hour", Pass = "none"),
  columns_to_remove = c("Serial", "FieldID", "LoadID")
)
}

}
\seealso{
\code{\link[=anonymize_coordinates]{anonymize_coordinates()}} pour l'anonymisation spatiale
}
