% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{render_pdf_report}
\alias{render_pdf_report}
\title{Render report to PDF using Chrome}
\usage{
render_pdf_report(html_file, pdf_file, verbose = TRUE)
}
\arguments{
\item{html_file}{Path to HTML file}

\item{pdf_file}{Path for output PDF file}

\item{verbose}{Print progress messages}
}
\value{
Path to the PDF file
}
\description{
Helper function to convert HTML to PDF using Chrome browser
}
