% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_diagnostics.R
\name{save_raster}
\alias{save_raster}
\title{Sauvegarder un raster en fichier}
\usage{
save_raster(raster, file_path, format = "tif")
}
\arguments{
\item{raster}{Objet SpatRaster}

\item{file_path}{Chemin du fichier de sortie}

\item{format}{Format de sortie ("tif", "asc", "grd")}
}
\description{
Sauvegarder un raster en fichier
}
