% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_jd_to_polygons}
\alias{read_jd_to_polygons}
\title{Lire des donnees John Deere et convertir en polygones metriques}
\usage{
read_jd_to_polygons(zip_path, field_name)
}
\arguments{
\item{zip_path}{Chemin vers le fichier ZIP}

\item{field_name}{Nom du champ dans le fichier ZIP}
}
\value{
Objet SF avec polygones et toutes les colonnes preservees
}
\description{
Lit un fichier ZIP John Deere (rendement, semis, vitesse, etc.) et cree des polygones.
Utilise les unites du JSON pour convertir en metrique.
}
